+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Zfs,rollback via restore: Fixed bug where instead of a rollback LINSTOR creates empty ZVOL
 - DRBD, rollback: Re-adjust (up) DRBD after rollback is finished (GitHub #464)
+- StorPool: Remote storage pools are now properly sent to satellites (not just local)
 
 ## [1.32.2] - 2025-10-01